+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/stack-backend
 
+## 2.8.34
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.34
+
+## 2.8.33
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.33
+
 ## 2.8.32
 
 ### Patch Changes